+++ HISTORY.rst
@@ -11,6 +11,8 @@ PlatformIO 3.0
 * Fixed incorrect wording when initializing/updating project
 * Fixed an issue with incorrect order for library dependencies ``CPPPATH``
   (`issue #1914 <https://github.com/platformio/platformio-core/issues/1914>`_)
+* Fixed an issue when Library Dependency Finder (LDF) does not handle project `src_filter <http://docs.platformio.org/page/projectconf/section_env_build.html#src-filter>`__
+  (`issue #1905 <https://github.com/platformio/platformio-core/issues/1905>`_)
 
 3.6.1 (2018-10-29)
 ~~~~~~~~~~~~~~~~~~